Switch to OpenCV 4.11.0 and update build system
Updated to use OpenCV version 4.11.0 by modifying the cache
key and git clone branches for both the `opencv` and
`opencv_contrib` repositories. Updated the build process
for OpenCVSharp Native by switching the `cmake` generator
from "Ninja" to "Visual Studio 17 2022" for ARM64.
Additionally, added a line to set the directory to the
GitHub workspace for better clarity.