Add GitHub Actions for ARM64 OpenCVSharp build
Introduce a new GitHub Actions workflow to automate the
building of OpenCVSharp for the ARM64 architecture on Windows.
This workflow triggers on pushes, pull requests, and manual
dispatch events for the main and master branches. It sets up
the environment, including vcpkg for Tesseract ARM64, and
utilizes MSBuild tools. Caching is implemented to optimize
subsequent builds. The workflow also uploads the built
artifacts for later retrieval.